requirements.txt
RPi.Gpio
Adafruit_DHT
